Split uniform wrapper module
