+++ source/framework/python/src/submodule.cpp
@@ -82,7 +82,8 @@ namespace lue::framework {
     void bind_locality_id(pybind11::module& module);
     void bind_normal(pybind11::module& module);
     void bind_timestamp(pybind11::module& module);
-    void bind_uniform(pybind11::module& module);
+    void bind_uniform1(pybind11::module& module);
+    void bind_uniform2(pybind11::module& module);
 
     void bind_read_array(pybind11::module& module);
     void bind_write_array(pybind11::module& module);
@@ -163,7 +164,8 @@ namespace lue::framework {
         }
 
         bind_normal(submodule);
-        bind_uniform(submodule);
+        bind_uniform1(submodule);
+        bind_uniform2(submodule);
 
         bind_model(submodule);
         bind_progressor(submodule);
